refactor: Consolidate environment variables and constants into a single file and update documentation.
