+++ deno.jsonc
@@ -24,7 +24,8 @@
     "mdast": "npm:mdast@^3",
     "turndown": "npm:turndown@^7.2.0",
     "unist-util-visit": "npm:unist-util-visit@^5.0.0",
-    "@std/dotenv": "jsr:@std/dotenv@^0.225.2"
+    "@std/dotenv": "jsr:@std/dotenv@^0.225.2",
+    "@bott/constants": "./constants.ts"
   },
   "tasks": {
     "reset": "rm -rf ./fs_root",
