+++ libraries/gemini/attachments/movie.ts
@@ -13,7 +13,7 @@ import { type GenerateVideosOperation, PersonGeneration } from "@google/genai";
 import { decodeBase64 } from "@std/encoding";
 
 import { BottAttachmentType } from "@bott/model";
-import { MOVIE_MODEL } from "../constants.ts";
+import { MOVIE_MODEL } from "@bott/constants";
 
 import _gemini from "../client.ts";
 import type { BottAttachmentDataGenerator } from "./types.ts";
